+++ packages/sources/coinpaprika/src/endpoint/crypto-single.ts
@@ -44,6 +44,40 @@ export const inputParameters: InputParameters = {
     type: 'string',
   },
 }
+export interface ResponseSchema {
+  id: string
+  name: string
+  symbol: string
+  rank: number
+  circulating_supply: number
+  total_supply: number
+  max_supply: number
+  beta_value: number
+  first_data_at: string
+  last_updated: string
+  quotes: {
+    [key: string]: {
+      price: number
+      volume_24h: number
+      volume_24h_change_24h: number
+      market_cap: number
+      market_cap_change_24h: number
+      percent_change_15m: number
+      percent_change_30m: number
+      percent_change_1h: number
+      percent_change_6h: number
+      percent_change_12h: number
+      percent_change_24h: number
+      percent_change_7d: number
+      percent_change_30d: number
+      percent_change_1y: number
+      ath_price: number
+      ath_date: string
+      percent_from_price_ath: number
+    }
+  }
+  cost?: number
+}
 
 export const execute: ExecuteWithConfig<Config> = async (request, context, config) => {
   const validator = new Validator(request, inputParameters)
@@ -74,9 +108,9 @@ export const execute: ExecuteWithConfig<Config> = async (request, context, confi
     params,
   }
 
-  const response = await Requester.request(options)
-  response.data.result = Requester.validateResultNumber(response.data, resultPath)
+  const response = await Requester.request<ResponseSchema>(options)
+  const result = Requester.validateResultNumber(response.data, resultPath)
   response.data.cost = 2
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }