+++ packages/composites/linear-finance/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/linear-finance-adapter
 
+## 1.1.18
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.0.21
+
+## 1.1.17
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.0.20
+
+## 1.1.16
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.0.19
+
 ## 1.1.15
 
 ### Patch Changes