+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.0.25
+
+### Patch Changes
+
+- @chainlink/ea@1.0.25
+
+## 1.0.24
+
+### Patch Changes
+
+- @chainlink/ea@1.0.24
+
+## 1.0.23
+
+### Patch Changes
+
+- @chainlink/ea@1.0.23
+
 ## 1.0.22
 
 ### Patch Changes