+++ packages/core/legos/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @chainlink/ea
 
+## 1.0.25
+
+### Patch Changes
+
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [cd01b0c5]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+- Updated dependencies [4d3c0a79]
+  - @chainlink/ada-balance-adapter@2.0.0
+  - @chainlink/tiingo-adapter@1.4.0
+  - @chainlink/chain-reserve-wallet-adapter@2.0.0
+  - @chainlink/renvm-address-set-adapter@1.1.0
+  - @chainlink/gemini-adapter@2.0.0
+  - @chainlink/lotus-adapter@2.0.0
+  - @chainlink/wbtc-address-set-adapter@1.1.0
+
+## 1.0.24
+
+### Patch Changes
+
+- Updated dependencies [56be0406]
+- Updated dependencies [e6700270]
+- Updated dependencies [d1c5127a]
+  - @chainlink/tiingo-adapter@1.3.0
+  - @chainlink/ncfx-adapter@1.1.0
+
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [19381a8d]
+- Updated dependencies [ab4f8eeb]
+  - @chainlink/cfbenchmarks-adapter@1.1.0
+  - @chainlink/ncfx-adapter@1.0.14
+  - @chainlink/nikkei-adapter@1.1.7
+  - @chainlink/nomics-adapter@1.0.14
+
 ## 1.0.22
 
 ### Patch Changes